Save a program's random numbers and count its console in a long
Random drew from a java.util.Random that the save never wrote down, so a
program put away and brought back started its sequence over and drew the
same numbers again. Its numbers now come from a generator of its own
(SplitMix64) whose one long of state goes into the save, and Random.Seed
starts it from that seed. A given seed therefore draws different numbers
than it did before.

The count of lines a program has written was an int, which a chatty
program that stays up would wrap. It is a long now, through the language
contract, the save and the terminal that shows what is new.

+++ computers/src/main/java/dev/jstech/computers/machine/CannonProgram.java
@@ -69,7 +69,7 @@ final class CannonProgram implements ILanguageProcess {
     }
 
     @Override
-    public int written() {
+    public long written() {
         return this.process.written();
     }
